Processamento de fila idempotente

+++ src/Backend/MyRecipeBook.Application/UseCases/User/Delete/Delete/DeleteUserAccountUseCase.cs
@@ -5,15 +5,18 @@ using MyRecipeBook.Domain.Services.Storage;
 namespace MyRecipeBook.Application.UseCases.User.Delete.Delete;
 public class DeleteUserAccountUseCase : IDeleteUserAccountUseCase
 {
+    private readonly IUserReadOnlyRepository _userReadOnlyRepository;
     private readonly IUserDeleteOnlyRepository _repository;
     private readonly IUnitOfWork _unitOfWork;
     private readonly IBlobStorageService _blobStorageService;
 
     public DeleteUserAccountUseCase(
+        IUserReadOnlyRepository userReadOnlyRepository,
         IUserDeleteOnlyRepository repository,
         IBlobStorageService blobStorageService,
         IUnitOfWork unitOfWork)
     {
+        _userReadOnlyRepository = userReadOnlyRepository;
         _blobStorageService = blobStorageService;
         _repository = repository;
         _unitOfWork = unitOfWork;
@@ -21,6 +24,10 @@ public class DeleteUserAccountUseCase : IDeleteUserAccountUseCase
 
     public async Task Execute(Guid userIdentifier)
     {
+        var userExists = await _userReadOnlyRepository.ExistActiveUserWithIdentifier(userIdentifier);
+        if (userExists.Equals(false))
+            return;
+
         await _blobStorageService.DeleteContainer(userIdentifier);
 
         await _repository.DeleteAccount(userIdentifier);
